Add regex example to "Match detection for URIs" that is not broken (#158)
* Add regex example to "Match detection for URIs" that is not broken

I find it important that examples actually follow best practices. The
current regex one is not. `^https://.*google\.com$` is an improper regex
(as already pointed out because it also matches `malicious-site.com`)
that is only there to show the weaknesses of regular expressions for
this use case. I find such an example very good for this purpose but
there should also be a "good example" that complements it.

I found such a "good example" that I hope is more useful and has no
unwanted loopholes.

* Change wording from bad/good to unsafe/safe in regex match detection

* Update uri-match-detection.md

Co-authored-by: fred_the_tech_writer <69817454+fschillingeriv@users.noreply.github.com>